+++ src/utils/information-extractor.js
@@ -1,10 +1,11 @@
 export const getVideoId = () => {
-  return new URLSearchParams(window.parent.top.location.search).get('v')
+  const searchParams = new URLSearchParams(window.location.search)
+  return searchParams.get('v') || new URLSearchParams(window.parent.top.location.search).get('v')
 }
 
-export const getChannelId = () => {
-  return window.parent.top.document
-    .querySelector('ytd-channel-name .yt-simple-endpoint')
-    .href.split('/')
-    .pop()
-}
+// export const getChannelId = () => {
+//   return window.parent.top.document
+//     .querySelector('ytd-channel-name .yt-simple-endpoint')
+//     .href.split('/')
+//     .pop()
+// }