+++ webpack.config.js
@@ -1,18 +1,18 @@
-const webpack = require('webpack');
-const ejs = require('ejs');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const CopyPlugin = require('copy-webpack-plugin');
-const ExtensionReloader = require('webpack-extension-reloader');
-const { VueLoaderPlugin } = require('vue-loader');
-const { version } = require('./package.json');
+const webpack = require('webpack')
+const ejs = require('ejs')
+const MiniCssExtractPlugin = require('mini-css-extract-plugin')
+const CopyPlugin = require('copy-webpack-plugin')
+const ExtensionReloader = require('webpack-extension-reloader')
+const { VueLoaderPlugin } = require('vue-loader')
+const { version } = require('./package.json')
 const path = require('path')
 
 const config = {
   mode: process.env.NODE_ENV,
   context: __dirname + '/src',
   entry: {
-    'background': './background.js',
-    'content': './content.js',
+    background: './background.js',
+    content: './content.js',
     'content-style': './content-style.scss',
     'popup/popup': './popup/popup.js',
     'options/options': './options/options.js',
@@ -24,7 +24,7 @@ const config = {
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src/'),
-      vue: 'vue/dist/vue.esm.js'
+      vue: 'vue/dist/vue.esm.js',
     },
     extensions: ['.js', '.vue', '.ts'],
   },
@@ -76,6 +76,20 @@ const config = {
           esModule: false,
         },
       },
+      {
+        test: /\.md$/,
+        use: [
+          {
+            loader: 'html-loader',
+          },
+          {
+            loader: 'markdown-loader',
+            options: {
+              /* your options here */
+            },
+          },
+        ],
+      },
     ],
   },
   plugins: [
@@ -93,20 +107,20 @@ const config = {
       {
         from: 'manifest.json',
         to: 'manifest.json',
-        transform: (content) => {
-          const jsonContent = JSON.parse(content);
-          jsonContent.version = version;
+        transform: content => {
+          const jsonContent = JSON.parse(content)
+          jsonContent.version = version
 
           if (config.mode === 'development') {
-            jsonContent['content_security_policy'] = "script-src 'self' 'unsafe-eval'; object-src 'self'";
+            jsonContent['content_security_policy'] = "script-src 'self' 'unsafe-eval'; object-src 'self'"
           }
 
-          return JSON.stringify(jsonContent, null, 2);
+          return JSON.stringify(jsonContent, null, 2)
         },
       },
     ]),
   ],
-};
+}
 
 if (config.mode === 'production') {
   config.plugins = (config.plugins || []).concat([
@@ -115,7 +129,7 @@ if (config.mode === 'production') {
         NODE_ENV: '"production"',
       },
     }),
-  ]);
+  ])
 }
 
 if (process.env.HMR === 'true') {
@@ -123,13 +137,13 @@ if (process.env.HMR === 'true') {
     new ExtensionReloader({
       manifest: __dirname + '/src/manifest.json',
     }),
-  ]);
+  ])
 }
 
 function transformHtml(content) {
   return ejs.render(content.toString(), {
     ...process.env,
-  });
+  })
 }
 
-module.exports = config;
+module.exports = config