+++ changelog.md
@@ -1,4 +1,21 @@
-# 1.2.0
+# ytcFilter Changelog
+
+## 1.3.0
+
+Changes:
+
+- Autoscroll is triggered only if the captured chat is at the bottom
+- Redesign notification
+
+New features:
+
+- Filter by moderator
+- Filter by owner
+- Resizing option
+- Support for embedded chat outside of Youtube
+- Update notification
+
+## 1.2.0
 
 Due to an unfortunate coincidence, the extension is now named ytcFilter.
 
@@ -7,18 +24,18 @@ New features:
 - Scroll to top/bottom buttons
 - Option to set a default profile which is automatically applied to new livechat
 
-# 1.1.2
+## 1.1.2
 
 Code cleaning
 
-# 1.1.1
+## 1.1.1
 
 Bugfix:
 
 - Case sensitive didn't work orz
 - Filter removal didn't save to storage
 
-# 1.1.0
+## 1.1.0
 
 New features:
 
@@ -32,6 +49,6 @@ Known issues:
 
 - Duplication will happen if the extension is used both on the live stream and VOD
 
-# 1.0.0
+## 1.0.0
 
 Initial release